feat: add delete note

+++ src/note/note.service.ts
@@ -1,6 +1,15 @@
-import { Injectable } from '@nestjs/common';
+import {
+  HttpException,
+  HttpStatus,
+  Injectable,
+  StreamableFile,
+} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
+import { createReadStream, existsSync } from 'fs';
+import { join, extname } from 'path';
 import { Repository } from 'typeorm';
+import * as fs from 'fs';
+
 import { CreateNoteDto } from './dto/create-note.dto';
 import { UpdateNoteDto } from './dto/update-note.dto';
 import { Note } from './entities/note.entity';
@@ -13,27 +22,53 @@ export class NoteService {
   ) {}
 
   async create(createNoteDto: CreateNoteDto, file: Express.Multer.File) {
-    //console.log(createNoteDto);
-    console.log(file);
+    const note = new Note(
+      createNoteDto.name,
+      createNoteDto.message,
+      file.filename,
+      file.originalname,
+    );
+
+    return await this.noteRepository.save(note);
   }
 
-  findAll() {
-    return this.noteRepository.find();
+  async findOne(id: string) {
+    const note = await this.noteRepository.findOne(id);
+    if (!note) {
+      throw new HttpException('File not found!', HttpStatus.NOT_FOUND);
+    }
+
+    const filePath = join(process.cwd(), `/uploads/${note.fileId}`);
+
+    if (existsSync(filePath)) {
+      const stream = createReadStream(filePath);
+      return new StreamableFile(stream);
+    }
+    throw new HttpException('File was not founded!', HttpStatus.NOT_FOUND);
   }
 
-  findOne(id: string) {
-    return this.noteRepository.findOne(id);
-  }
-
-  async update(id: string, updateNoteDto: UpdateNoteDto) {
+  async update(
+    id: string,
+    updateNoteDto: UpdateNoteDto,
+    file: Express.Multer.File,
+  ) {
     return `This action updates a #${id} note`;
   }
 
   async remove(id: string) {
     const delNote = await this.noteRepository.findOne(id);
     if (!delNote) {
-      return false;
+      throw new HttpException('ID not found!', HttpStatus.NOT_FOUND);
     }
+    const filepath = join(process.cwd(), `/uploads/${delNote.fileId}`);
+
+    if (!existsSync(filepath)) {
+      throw new HttpException('File was not founded!', HttpStatus.NOT_FOUND);
+    }
+    fs.unlink(filepath, (err) => {
+      if (err) throw new HttpException('File not found!', HttpStatus.NOT_FOUND);
+    });
+
     return this.noteRepository.remove(delNote);
   }
 }
